requirements.txt
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-sqlite==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.0
pillow-heif==0.12.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
